Bury spaceguns (sorry, maybe revive later!)
